chore: tasks polling moved to component

+++ src/app/features/users/facade/users-facade.ts
@@ -1,13 +1,15 @@
-import { Injectable, OnDestroy } from "@angular/core";
+import { inject, Injectable, OnDestroy } from "@angular/core";
 import { UsersApiService } from "../api/users-api-service";
 import { environment } from "../../../../environments/environment";
-import { BehaviorSubject, interval, startWith, Subject, switchMap, takeUntil } from "rxjs";
+import { BehaviorSubject, interval, Observable, startWith, Subject, switchMap, takeUntil, throwError } from "rxjs";
 import { UserModel } from "../models/users-model";
+import { TasksApiService } from "../../tasks/api/tasks-api-service";
 
 @Injectable({
   providedIn: 'root'
 })
 export class UsersFacade implements OnDestroy {
+    private readonly tasksService = inject(TasksApiService);
     private readonly userPollingFreq = environment.userPollingFrequency;
     allUsersPolling = new BehaviorSubject<UserModel[]>([]);
     allUsers$ = this.allUsersPolling.asObservable();
@@ -34,8 +36,15 @@ export class UsersFacade implements OnDestroy {
         return this.usersApiService.createUser(user);
     }
 
-    deleteUser(id: number) {
-        return this.usersApiService.deleteUser(id);
+    deleteUser(id: number): Observable<void> {
+        return this.tasksService.getAllTasksByUserId(id).pipe(
+            switchMap((tasks) => {
+                if (tasks.length > 0) {
+                    return throwError(() => new Error('Cannot delete user with assigned tasks'));
+                }
+                return this.usersApiService.deleteUser(id);
+            })
+        );
     }
 
     ngOnDestroy(): void {
